requirements.txt
requests==2.6.0
pycrypto==2.6.1
